argon2, bcrypt: reworked how backends are managed.
These two classes have very complex backends, with lots of tests,
and backends that override multiple methods.  Instead of awkwardly
patching a bunch of things on the class when backend is loaded,
redid these to use new strategy:

A base class (e.g. _BCryptCommon) is defined with all the common code,
and subclasses are defined for each backend (e.g. _PyBCryptBackend).
.set_backend() now dynamically modifies the bases of the hash class
to include the appropriate backend-specific subclass; allowing each
backend to easily override any methods it needs to.

Changes:

* Abstracted out the cls.__bases__ modifying code into SubclassBackendMixin

* Argon2 class split into multiple backend classes;
  removes need for monkeypatching & redecorate_method() hack

* Bcrypt class split into multiple backend classes.

* Refactored bcrypt backend compatibility workaround tests:

        - now initialized once for each backend, results stored in mixin class.
          should make UTs switch backends MUCH faster

        - tests for known bugs against all ident variants,
          in case behaior is different between them.

        - bug checks test for incorrect AND correct hashes,
          to detect unexpected behaviors

        - wraparound bug still handled for 2a, but now fatal for 2y/2by

        - 2b ident now preferred as fallback if supported